@@ -1,12 +1,12 @@
-import { useState } from 'react'
 import './App.css'
+import Layout from './components/layout/Layout'
 
 function App() {
 
   return (
-    <>
-      App
-    </>
+    <Layout>
+      Children
+    </Layout>
   )
 }
 
